On Android phones with higher density and newer screens the buttons are so small they cannot be touched at all.
In testing this with an actual child on a S8+ (1440x2960 with 529ppi) and a S22+ (1080x2340 with 393ppi) the sweet spot appears to be buttons tha take up no less than 4.5% of the screen width. Below this they appear to be unusuable, and above this there's too much scroll of the toolbars.

 Button scale is incremented by 25% on each pass until the condition is met.