Changing Android save paths to now default to the 'external' dir from the device itself.
On most devices this isn't actually on the SD card, as most no longer support that, and instead return '/storage/emulated/0/Android/data/org.tuxpaint/'

This directory is not accessible in API31 though from a file explorer, so the export path must be moved to the normal /storage/emulated/0/Pictures/TuxPaint position where most apps store pictures. This directory is only used when SDL_AndroidGetExternalStoragePath returns the internal path and not external.

The default of /mnt/sdcard/TuxPaint is removed entirely from the default config as it doesn't work for most devices.